seleccionar y mostar una imagen local

+++ src/modules/daybook/views/EntryView.vue
@@ -8,6 +8,7 @@
                 <span class="mx-2 fs-4 fw-light">{{yearDay}}</span>
             </div>
             <div>
+                <input type="file" @change="onSelectedImage"/>
                 <button v-if="entry.id" @click="onDeleteEntry" class="btn btn-danger mx-2">
                     <i class="fa fa-trash-alt"></i>                
                     Delete
@@ -25,7 +26,10 @@
                 placeholder="¿Que sucedio hoy?"
             ></textarea>
         </div>
-        <img src="https://www.hindustantimes.com/ht-img/img/2023/04/07/1600x900/jujutsu-kaisen_1680853025638_1680853037810_1680853037810.jpeg" alt="entry-picture" class="img-thumbnail"/>
+        
+        <!-- <img src="https://www.hindustantimes.com/ht-img/img/2023/04/07/1600x900/jujutsu-kaisen_1680853025638_1680853037810_1680853037810.jpeg" alt="entry-picture" class="img-thumbnail"/> -->
+        
+        <img v-if="localImage" :src="localImage" alt="entry-picture" class="img-thumbnail"/>
     </template>
     <FabComponent 
     icon="fa-save"
@@ -36,6 +40,8 @@
 <script>
 import { defineAsyncComponent } from 'vue'
 import { mapGetters, mapActions } from 'vuex';
+import Swal from 'sweetalert2'
+
 import getDayMonthYear from '../helpers/getDayMonthYear';
 export default {
     props: {
@@ -50,7 +56,9 @@ export default {
     },
     data() {
         return {
-            entry: null
+            entry: null,
+            localImage: null,
+            file: null
         }
     },
     computed: {
@@ -86,8 +94,16 @@ export default {
         /**
          * This method is dispached when the fab button is click activating the on:event
          */
-        ...mapActions('journal', ['updateEntry', 'createEntry','deleteEntry']),
+        ...mapActions('journal', ['updateEntry', 'createEntry', 'deleteEntry']),
         async saveEntry() {
+
+            new Swal({
+                title: 'Please wait',
+                allowOutsideClick: false
+            });
+
+            Swal.showLoading()
+
             if (this.entry.id) {
                 // Call update entry action
                 await this.updateEntry(this.entry);
@@ -95,16 +111,55 @@ export default {
                 // Call create entry action
                 const id = await this.createEntry(this.entry)
                 // Redirect to a entry view
-                this.$router.push({name: 'entry', params:{id} })
+                this.$router.push({ name: 'entry', params: { id } })
+            }
+
+            Swal.fire('Saved', 'The register was saved', 'success')
+        },
+        async onDeleteEntry() {
+            const { isConfirmed } = await Swal.fire({
+                title: ' Are you sure?',
+                text: 'If you delete this content, you can not recovery it after',
+                showDenyButton: true,
+                confirmButtonText: 'Yes, I am sure!'
+
+            });
+
+            if (isConfirmed) {
+                new Swal({
+                    title: 'Please wait',
+                    allowOutsideClick: false
+                });
+
+                Swal.showLoading()
+                await this.deleteEntry(this.entry.id)
+                //Redirect the user to entry 
+                this.$router.push({ name: 'no-entry' })
+                Swal.fire('Deleted','', 'success')
             }
         },
-        async onDeleteEntry(){
-            this.deleteEntry(this.entry.id)
-            //Redirect the user to entry 
-            this.$router.push({name:'no-entry'})
+        onSelectedImage(event){
+            const file = event.target.files[0];
+
+            // Validate is a image was sended
+            if(!file){
+                // If the send of the image was cnacel the picture will be null
+                this.localImage = null;
+                this.file = null
+                return
+            }
+            this.file = file
+
+            // This is a javascript built-in object to manage files
+            const fr = new FileReader()
+            // 1. wait for the element charge the save he result in the localImage
+            // 2. Use the method readAsDataURL() and send the file to save
+            fr.onload = () => this.localImage  =  fr.result
+            fr.readAsDataURL( file )
+
         }
     },
-    created(){
+    created() {
 
         this.loadEntry();
     },
